requirements.txt
absl-py==1.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==7.1.2
colorama==0.4.5
cv2PySide6==3.1.0
cycler==0.11.0
decorator==5.1.1
executing==0.10.0
fonttools==4.37.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
importlib-metadata==4.12.0
jedi==0.18.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mediapipe==0.8.10.1
micropython-esp32-stubs==1.19.1.post3
numpy==1.23.2
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.20.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
pyqt6-plugins==6.1.0.2.2
PyQt6-Qt6==6.3.1
PyQt6-sip==13.4.0
pyqt6-tools==6.1.0.3.2
pyqtgraph==0.12.4
PySide6==6.3.1
PySide6-Addons==6.3.1
PySide6-Essentials==6.3.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
qimage2ndarray==1.9.0
qt6-applications==6.1.0.2.2
qt6-tools==6.1.0.1.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
shiboken6==6.3.1
six==1.16.0
stack-data==0.4.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.8.1
